Add basic URL support, and refactor
